Updated controllers and transfomers

+++ app/Http/Controllers/MaintenanceTypesController.php
@@ -29,6 +29,7 @@ class MaintenanceTypesController extends Controller
 
         $type = new MaintenanceType;
         $type->name = $request->input('name');
+        $type->tag_color = $request->input('tag_color');
         $type->created_by = auth()->id();
 
         if ($type->save()) {
@@ -37,6 +38,13 @@ class MaintenanceTypesController extends Controller
         }
 
         return redirect()->back()->withInput()->withErrors($type->getErrors());
+    }
+
+    public function show(MaintenanceType $maintenanceType): View
+    {
+        $this->authorize('view', $maintenanceType);
+
+        return view('maintenance-types.view')->with('item', $maintenanceType);
     }
 
     public function edit(MaintenanceType $maintenanceType): View
@@ -51,6 +59,7 @@ class MaintenanceTypesController extends Controller
         $this->authorize('update', $maintenanceType);
 
         $maintenanceType->name = $request->input('name');
+        $maintenanceType->tag_color = $request->input('tag_color');
 
         if ($maintenanceType->save()) {
             return redirect()->route('maintenance-types.index')
@@ -63,6 +72,18 @@ class MaintenanceTypesController extends Controller
     public function destroy(MaintenanceType $maintenanceType): RedirectResponse
     {
         $this->authorize('delete', $maintenanceType);
+
+        // Same guard as the API destroy: refuse when any maintenance row
+        // still references this type so we don't strand orphaned
+        // maintenance_type_id values on a soft-deleted parent.
+        if (! $maintenanceType->isDeletable()) {
+            return redirect()->route('maintenance-types.index')
+                ->with('error', trans('general.bulk_delete_associations.assoc_maintenances', [
+                    'item_name' => $maintenanceType->name,
+                    'maintenance_count' => $maintenanceType->maintenances()->count(),
+                    'item' => trans('admin/maintenance_types/general.maintenance_type'),
+                ]));
+        }
 
         $maintenanceType->delete();
